+++ packages/pigeon/test/dart/proxy_api_test.dart
@@ -472,7 +472,13 @@ void main() {
         expect(
           collapsedCode,
           contains(
-            r'pigeonVar_channel .send(<Object?>[pigeonVar_instanceIdentifier])',
+            'pigeonVar_sendFuture = pigeonVar_channel.send(<Object?>[pigeonVar_instanceIdentifier]);',
+          ),
+        );
+        expect(
+          collapsedCode,
+          contains(
+            '() async { final List<Object?>? pigeonVar_replyList = await pigeonVar_sendFuture as List<Object?>?;',
           ),
         );
       });
@@ -903,11 +909,12 @@ void main() {
         expect(
           collapsedCode,
           contains(
-            r'await pigeonVar_channel.send(<Object?>[ this, validType, '
+            r'pigeonVar_channel.send(<Object?>[ this, validType, '
             r'enumType, proxyApiType, nullableValidType, '
             r'nullableEnumType, nullableProxyApiType ])',
           ),
         );
+        expect(code, contains('await pigeonVar_sendFuture'));
       });
 
       test('static method', () {
@@ -951,8 +958,9 @@ void main() {
         );
         expect(
           collapsedCode,
-          contains(r'await pigeonVar_channel.send(null)'),
+          contains(r'pigeonVar_channel.send(null)'),
         );
+        expect(code, contains('await pigeonVar_sendFuture'));
       });
     });